added button for scriptlibrary table, because it shows off the edge of the window.  So I added a button called tips and connect the tips to that button.
added a message when the NNPAPI is opening and change colors of button when finished because it takes a while for the process to finish and this give clear indication of when to continue.

added a logical checkbox table in emgRecordProcessedTab for setting which hand patterns will have the lock function disabled.   I may want to duplicate this elsewhere, like in the  actionPatternMap application,  but I had so many other parameters being set here, and already had the uint8 version of the parameter there already.

git-svn-id: https://jdc-consulting.unfuddle.com/svn/jdc-consulting_matlabapi@218 03bc5d11-0c36-4e61-bad9-4efbbe88aa87